Make the display interactive
Create a function 'operate()' that uses the arguments 'a', 'b', 'c'
in order to run a function requested by the user
If argument 'b' is a symbol that represents a mathematical operation,
run the corresponding function.

Add the user interface with buttons and a div that displays the result
in order for the user to use the buttons instead of typing out the
function in the console

Make numbers appear on the display when the user clicks the numbered
buttons using an eventListener and a variable that stores the value
of a clicked button

Add the class 'number' to the numbers with the numeric value to select
them as a value for the 'buttons' variable.