The program now displays full numbers
Previously the program displayed only the number or the symbol that was
clicked the last, e.g. when typing 999, the only number that was displayed
was 9, even though the number was made up of 3 digits. The improvement
is needed in order to insure that the user sees the number that they
want to perform an operation on.